the led now takes the data from the buffer as it should added setColorInRange both setStripColor and setColorInRange can receive Color too now.
